+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Support nextLink in list Container with Management plane API
+    -  Get-AzRmStorageContainer
 * Updated example in reference documentation for `Get-AzStorageAccountKey`
 * In upload/Downalod Azure File,support perserve the source File SMB properties (File Attributtes, File Creation Time, File Last Write Time) in the destination file
     -  Set-AzStorageFileContent